Update syntax file.
This update is taken directly from the vim source code repository. Apparently there has been two channels of development for the yaml.vim syntax file with no communication between the two groups. Since the one more closely associated with vim development appears to be the more mature version, that is the one being submitted here. But in the future it would be nice if this repo were to become the official one for all improvements going forward. 

Note, if anything feature was lost from the previous version, please submit a new pull request to re-add it.